Create README.md
The way i develop the program in java is that i got all the content as a string variable. With this it was easy to write and read to the file. For the txt files it was easy to get the content as a string without using any external library. But for the doc file, i couldn't just read the content form the doc, so that where apache poi came in. with this external library i was able to use some of its function to extract the content as a string. Writing to the files were as easy as reading it. I used  apache poi to write the content of string to a doc file  and i use BufferedWriter to write contents of string to txt file. Also having convert everything into string made easy to develop the count word function where it would only take a string as an argument.